page_order_auto/requirements.txt
requests>=2.25.1
pytest>=6.2.5
pytest-html>=3.1.1
allure-pytest>=2.9.45
PyYAML>=6.0